Handle haptic pulses in index.js
